finished moving, working on rom but have to make superstructureElement interface first
