+++ src/main/kotlin/frc/team449/Robot.kt
@@ -9,6 +9,7 @@ import frc.team449.subsystems.WebConnection
 import frc.team449.subsystems.drive.swerve.SwerveDrive
 import frc.team449.subsystems.drive.swerve.SwerveOrthogonalCommand
 import frc.team449.subsystems.light.Light.Companion.createLight
+import frc.team449.subsystems.superstructure.BuiltInTests
 import frc.team449.subsystems.superstructure.SuperstructureManager
 import frc.team449.subsystems.superstructure.SuperstructureManager.Companion.createSuperstructureManager
 import frc.team449.subsystems.superstructure.elevator.Elevator
@@ -61,6 +62,8 @@ class Robot : RobotBase() {
   val superstructureManager: SuperstructureManager = createSuperstructureManager(this)
 
   val light = createLight()
+
+  val tester = BuiltInTests(this)
 
   val webCom = WebConnection(this)
 }
