+++ src/main/kotlin/frc/team449/subsystems/superstructure/elevator/ElevatorConstants.kt
@@ -60,16 +60,7 @@ object ElevatorConstants {
   const val KD = 0.00015
 
   // Motion Magic
-  const val CRUISE_VEL_VALUE = 2.325 // max theoretical 2.66 m/s, max practical ? m/s from feedforward
-  const val MAX_ACCEL_VALUE = 7.5 // should get to 15.0, max theoretical ? m/s/s at 4kg and no gravity
-  var CRUISE_VEL = CRUISE_VEL_VALUE
-  var MAX_ACCEL = MAX_ACCEL_VALUE
-
-  fun changeCruiseVel(newCruiseVel: Double) {
-    CRUISE_VEL = newCruiseVel
-  }
-  fun changeMaxAccel(newMaxAccel: Double) {
-    MAX_ACCEL = newMaxAccel
-  }
+  const val CRUISE_VEL = 2.325 // max theoretical 2.66 m/s, max practical ? m/s from feedforward
+  const val MAX_ACCEL = 7.5 // should get to 15.0, max theoretical ? m/s/s at 4kg and no gravity
 
 }
